addTableColumn UPDATE
Removed for loop for adding multiple columns and instead implemented where only one column could be added at a time - code simplicity and reduces the likelihood of errors.

Updated the if statement that checks for "String" to VARCHAR

Implemented a variable for nullSetting and auto sets the primary key column to "not null".